% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_int_as_char.R
\name{add_int_as_char}
\alias{add_int_as_char}
\title{A function to add integers as strings}
\usage{
add_int_as_char(a, b)
}
\arguments{
\item{a}{integers must be in character format}

\item{b}{integer must in in character format}
}
\description{
This function allows you to add two integers reprsented as characters
}
\examples{
add_int_as_char()
}
\keyword{add}
\keyword{large}
\keyword{numbers}
