% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_as_string.R
\name{fact_as_string}
\alias{fact_as_string}
\title{A function to find large factorials}
\usage{
fact_as_string(a)
}
\arguments{
\item{a}{the factorial you want to calculate}
}
\description{
This function allows you find the solutions of large factorials as a character
}
\examples{
fact_as_string()
}
\keyword{factorial}
\keyword{large}
