% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_eval.R
\name{lib_eval}
\alias{lib_eval}
\title{A function to simplify dependencies}
\usage{
lib_eval(package)
}
\arguments{
\item{package}{Package name which must be in a string}
}
\description{
This function allows you to call a package and install if not present
}
\examples{
lib_eval()
}
\keyword{install}
\keyword{library}
