% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_int_as_char.R
\name{mult_int_as_char}
\alias{mult_int_as_char}
\title{A function multiply integers represeted as characters}
\usage{
mult_int_as_char(a, b)
}
\arguments{
\item{a}{integers must be in character format}

\item{b}{integer must in in character format}
}
\description{
This function allows you multiply large numbers
}
\examples{
mult_int_as_char()
}
\keyword{large}
\keyword{multiply}
\keyword{numbers}
